combine different 'apps' into one
All basically related to Gym
