+++ back/app/urls.py
@@ -1,9 +1,7 @@
 from django.contrib import admin
 from django.urls import path, include
 
-from core.gym.urls import gym_router
-from core.gym_details.urls import gym_details_router
-from core.gym_leaderboard.urls import gym_leaderboard_router
+from core.gym.urls import gym_router, gym_details_router, gym_leaderboard_router
 
 
 urlpatterns = [
