--- back/core/gym_leaderboard/serializer.py
@@ -1,12 +1,0 @@
-from rest_framework import serializers
-from core.gym_leaderboard.models import GymLeaderboard
-
-
-class GymLeaderboardSerializer(serializers.HyperlinkedModelSerializer):
-    class Meta:
-        model = GymLeaderboard
-        fields = [
-            'id',
-            'leaderboard_api_id',
-            'data'
-        ]
